Package: re-implement proxy support
Since we're not using superagent to get the package files,
superagent-proxy wasn't really taking effect. I've extracted the
proxy agent mapping and caching logic into the `proxy-agent` module,
so that we can use it directly with the `http`/`https` modules.

This fixes the --proxy and/or `http_proxy` env variables for realz